Tweak button fonts and colors a bit
