+++ src/vs/workbench/contrib/positronNotebook/browser/PositronNotebookHeader.css
@@ -1,0 +1,27 @@
+/*---------------------------------------------------------------------------------------------
+ *  Copyright (C) 2024 Posit Software, PBC. All rights reserved.
+ *--------------------------------------------------------------------------------------------*/
+
+.positron-notebook-header {
+	font-size: 12px;
+	color: var(--vscode-positronNotebook-deemphasized-foreground);
+	height: var(--vscode-positronNotebook-header-h);
+	gap: 2px;
+	padding-inline: calc(var(--vscode-positronNotebook-header-h) / 2);
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	border-bottom: 1px solid var(--vscode-positronNotebook-header-border);
+
+	.positron-notebook-header-divider {
+		height: calc(var(--vscode-positronNotebook-header-h) / 2);
+		width: 1px;
+		background-color: var(--vscode-positronNotebook-header-border);
+	}
+
+	button.action-button,
+	.codicon {
+		color: var(--vscode-positronNotebook-deemphasized-foreground);
+	}
+}
+
