+++ src/vs/workbench/contrib/positronNotebook/browser/utilityComponents/IconedButton.css
@@ -6,14 +6,17 @@
 	cursor: pointer;
 	display: flex;
 	gap: 6px;
+	font-size: 12px;
 	color: var(--vscode-positronNotebook-deemphasized-foreground);
 
 	.button-icon {
 		color: inherit;
 		font-size: 18px;
+		transform: translateY(0px);
 	}
 
 	.action-label {
+		margin-inline-end: 0.2rem;
 		cursor: pointer;
 	}
 }
